pyproject.toml
[build-system]
requires = ["setuptools>=69"]
build-backend = "setuptools.build_meta"

[project]
name = "belediye-atik-sistemi"
version = "0.1.0"
description = "Akilli sehir atik yonetimi ve geri donusum sistemi backend/API altyapisi."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.13",
    "fastapi>=0.115",
    "passlib[bcrypt]>=1.7",
    "psycopg[binary]>=3.2",
    "pydantic-settings>=2.6",
    "python-jose[cryptography]>=3.3",
    "python-multipart>=0.0.12",
    "sqlalchemy>=2.0",
    "uvicorn[standard]>=0.32",
]

[project.optional-dependencies]
dev = [
    "httpx>=0.27",
    "pytest>=8.3",
    "ruff>=0.8",
]

[tool.pytest.ini_options]
testpaths = ["backend/tests"]
pythonpath = ["."]

[tool.setuptools.packages.find]
include = ["backend*"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.ruff.lint.per-file-ignores]
"backend/alembic/versions/*.py" = ["E501", "I001"]
